Split server into subdomain-routed apps (oaf, calendar, tulips)
Route requests by subdomain using req.hostname, replacing the single
monolithic Express app with separate apps under server/apps/. Each app
has its own API routes, web SPA, and vite config. Locally, the oaf app
is the default fallback when no subdomain is present.
