+++ packages/oaf/eslint.config.ts
@@ -10,7 +10,11 @@ export default defineConfig(
     languageOptions: {
       parserOptions: {
         projectService: {
-          allowDefaultProject: ["eslint.config.ts", "vitest.config.ts"],
+          allowDefaultProject: [
+            "eslint.config.ts",
+            "vitest.config.ts",
+            "vite.*.config.ts",
+          ],
         },
         tsconfigRootDir: import.meta.dirname,
       },
