+++ packages/oaf/src/server/apps/calendar/web/App.tsx
@@ -1,0 +1,20 @@
+import { BrowserRouter, Route, Routes } from "react-router";
+
+import CalendarPage from "./pages/CalendarPage.js";
+
+export default function App() {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route
+          path="/"
+          element={
+            <div className="container">
+              <CalendarPage />
+            </div>
+          }
+        />
+      </Routes>
+    </BrowserRouter>
+  );
+}
