+++ packages/oaf/src/server/apps/calendar/web/components/BouncingEmoji.tsx
