+++ packages/oaf/src/server/apps/calendar/web/components/holidayEmoji.ts
