+++ packages/oaf/src/server/apps/calendar/web/public/comet.gif
