+++ packages/oaf/src/server/apps/calendar/web/public/favicon.png
