+++ packages/oaf/src/server/apps/calendar/web/public/rift.gif
