+++ packages/oaf/src/server/apps/calendar/web/public/vortex.gif
