+++ packages/oaf/src/server/apps/oaf/api/offers.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 
-import { dataOfLoathingClient } from "../../clients/dataOfLoathing.js";
-import { getAllOffersWithBuyers } from "../../clients/database.js";
+import { dataOfLoathingClient } from "../../../../clients/dataOfLoathing.js";
+import { getAllOffersWithBuyers } from "../../../../clients/database.js";
 
 export const offersRouter = Router();
