+++ packages/oaf/src/server/apps/oaf/api/pilot.ts
@@ -1,7 +1,7 @@
 import { messageLink } from "discord.js";
 import { Router } from "express";
 
-import { discordClient } from "../../clients/discord.js";
+import { discordClient } from "../../../../clients/discord.js";
 
 export const pilotRouter = Router();
