+++ packages/oaf/src/server/apps/oaf/web/index.html
@@ -5,12 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>O.A.F.</title>
     <link rel="icon" href="/favicon.png" type="image/png" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=UnifrakturMaguntia&display=swap"
-      rel="stylesheet"
-    />
     <link rel="stylesheet" href="/style.css" />
   </head>
   <body>
